requirements.txt
requests
openpyxl
pyyaml
pymysql

pytest
